실행전

+++ newdj/djsite/djapp/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 class Searchdb(models.Model):
     timedata = models.CharField(max_length=255)
-    distance = models.IntegerField()
+    distance = models.FloatField()
 
     class Meta:
         verbose_name_plural = "djapp"
